requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.1
h11==0.16.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
